Creates CameraPermissionsHelper to isolate permission logic into single class
