+++ app/src/main/java/com/willowtree/vocable/utils/FaceTrackingManager.kt
@@ -1,24 +1,13 @@
 package com.willowtree.vocable.utils
 
-import android.Manifest
-import android.app.Activity
 import android.app.ActivityManager
 import android.content.Context
-import android.content.Intent
-import android.content.pm.PackageManager
 import android.hardware.display.DisplayManager
-import android.net.Uri
-import android.provider.Settings
 import android.util.DisplayMetrics
 import android.view.Surface
 import android.view.View
 import android.widget.Toast
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContract
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityCompat
 import com.google.ar.core.ArCoreApk
 import com.willowtree.vocable.BuildConfig
 import com.willowtree.vocable.R
@@ -46,6 +35,7 @@ class FaceTrackingManager(
 
     val displayMetrics = DisplayMetrics()
 
+    private val cameraPermissionsHelper = CameraPermissionsHelper(activity)
     private lateinit var faceTrackingPointerUpdates: FaceTrackingPointerUpdates
 
     /**
@@ -58,11 +48,20 @@ class FaceTrackingManager(
         activity.windowManager.defaultDisplay.getMetrics(displayMetrics)
 
         if (BuildConfig.USE_HEAD_TRACKING && checkIsSupportedDevice()) {
+
+            cameraPermissionsHelper.permissionCallback = { isGranted ->
+                if (isGranted) {
+                    faceTrackingPermissions.enableFaceTracking()
+                } else {
+                    faceTrackingPermissions.disableFaceTracking()
+                }
+            }
+
             CoroutineScope(Dispatchers.Main).launch {
                 faceTrackingPermissions.permissionState.collect { headTrackingState ->
                     when (headTrackingState) {
                         IFaceTrackingPermissions.PermissionState.PermissionRequested -> {
-                            requestPermissions()
+                            cameraPermissionsHelper.requestPermissions()
                         }
 
                         IFaceTrackingPermissions.PermissionState.Enabled -> {
@@ -205,88 +204,4 @@ class FaceTrackingManager(
                 .commitAllowingStateLoss()
         }
     }
-
-    //region Permissions
-
-    private val permissionLauncher: ActivityResultLauncher<String> =
-        activity.registerForActivityResult(ActivityResultContracts.RequestPermission()) { isGranted: Boolean ->
-            if (isGranted) {
-                faceTrackingPermissions.enableFaceTracking()
-            } else {
-                faceTrackingPermissions.disableFaceTracking()
-                permissionSettingsDialog.show()
-            }
-        }
-
-    private val permissionResultLauncher: ActivityResultLauncher<String> =
-        activity.registerForActivityResult(object : ActivityResultContract<String, Boolean>() {
-            override fun createIntent(context: Context, input: String): Intent {
-                return Intent().apply {
-                    action = Settings.ACTION_APPLICATION_DETAILS_SETTINGS
-                    data = Uri.fromParts("package", activity.packageName, null)
-                }
-            }
-
-            override fun parseResult(resultCode: Int, intent: Intent?): Boolean {
-                return when (resultCode) {
-                    Activity.RESULT_OK, Activity.RESULT_CANCELED -> true
-                    else -> false
-                }
-            }
-        }) { result ->
-            if (result) {
-                permissionLauncher.launch(Manifest.permission.CAMERA)
-            }
-        }
-
-    private val permissionRationaleDialog by lazy {
-        AlertDialog.Builder(activity)
-            .setTitle(activity.getString(R.string.permissions_rationale_title))
-            .setPositiveButton(activity.getString(R.string.permissions_confirm)) { _, _ ->
-                permissionLauncher.launch(Manifest.permission.CAMERA)
-            }
-            .setNegativeButton(activity.getString(R.string.permissions_cancel)) { dialog, _ ->
-                faceTrackingPermissions.disableFaceTracking()
-                dialog.dismiss()
-            }
-            .setOnDismissListener { dialog ->
-                faceTrackingPermissions.disableFaceTracking()
-                dialog.dismiss()
-            }.create()
-    }
-
-    private val permissionSettingsDialog by lazy {
-        AlertDialog.Builder(activity)
-            .setTitle(activity.getString(R.string.permissions_missing_dialog_title))
-            .setMessage(activity.getString(R.string.permissions_missing_dialog_body))
-            .setPositiveButton(activity.getString(R.string.permissions_confirm)) { _, _ ->
-                permissionResultLauncher.launch(Manifest.permission.CAMERA)
-            }
-            .setNegativeButton(activity.getString(R.string.permissions_cancel)) { dialog, _ ->
-                faceTrackingPermissions.disableFaceTracking()
-                dialog.dismiss()
-            }
-            .create()
-    }
-
-    private fun requestPermissions() {
-
-        // Bypass check if we already have permission
-        if (ActivityCompat.checkSelfPermission(activity, Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED) {
-            faceTrackingPermissions.enableFaceTracking()
-            return
-        }
-
-        // Permission has been denied before, show rationale
-        if (ActivityCompat.shouldShowRequestPermissionRationale(activity, Manifest.permission.CAMERA)) {
-            permissionRationaleDialog.show()
-            return
-        }
-
-        // Ask for permissions.
-        permissionLauncher.launch(Manifest.permission.CAMERA)
-    }
-
-    //endregion Permissions
-
 }
